% Generated by build script
\name{hdx_dataset}
\alias{hdx_dataset}
\title{Get a specific HDX dataset by ID or name}
\usage{
hdx_dataset()
}
\arguments{
\item{id}{Dataset ID or name slug}
}
\value{tibble: id, name, title, organization, metadata_created,}
\description{Get a specific HDX dataset by ID or name}
