% Generated by build script
\name{hdx_organizations}
\alias{hdx_organizations}
\title{List HDX organizations}
\usage{
hdx_organizations()
}
\arguments{
\item{limit}{Max results (default 20)}
\item{offset}{Offset for pagination (default 0)}
}
\value{tibble: id, name, title, description, package_count}
\description{List HDX organizations}
