% Generated by build script
\name{hdx_search}
\alias{hdx_search}
\title{Search HDX datasets}
\usage{
hdx_search()
}
\arguments{
\item{query}{Search query (e.g. "refugees", "food security", "cholera")}
\item{rows}{Number of results (default 10)}
\item{start}{Offset for pagination (default 0)}
}
\value{tibble: id, name, title, organization, num_resources,}
\description{Search HDX datasets}
